kernel: vfs cleanup, some elf parsing, stalling for multitasking... 	modified:   Makefile 	new file:   initrd/test 	modified:   initrd/test.txt 	modified:   iso/boot/grub/grub.cfg 	modified:   kernel/cpu/cpuid.c 	modified:   kernel/cpu/tss.c 	modified:   kernel/entry.asm 	modified:   kernel/fs/devfs.c 	modified:   kernel/fs/devices/null.c 	modified:   kernel/fs/devices/port.c 	new file:   kernel/fs/devices/rand.c 	modified:   kernel/fs/rootfs.c 	modified:   kernel/fs/tarfs.c 	modified:   kernel/fs/vfs.c 	modified:   kernel/include/fs/devfs.h 	modified:   kernel/include/fs/fs.h 	modified:   kernel/include/fs/tarfs.h 	modified:   kernel/include/memory/kheap.h 	modified:   kernel/include/memory/vmm.h 	modified:   kernel/include/string.h 	modified:   kernel/include/sys/elf.h 	new file:   kernel/include/types.h 	modified:   kernel/kernel.c 	modified:   kernel/linker.ld 	modified:   kernel/memory/vmm.c 	modified:   kernel/sys/elf.c 	new file:   kernel/sys/usermode.asm 	modified:   kernel/utils/string.c 	renamed:    userland/user-link.ld -> userland/linker.ld

+++ kernel/fs/devices/null.c
@@ -1,6 +1,6 @@
 #include <fs/devfs.h>
 
-static size_t read_null(fs_node_t* node, uint32_t offset, size_t size, uint8_t* buf) {
+static ssize_t read_null(fs_node_t* node, off_t offset, size_t size, uint8_t* buf) {
     (void) node;
     (void) offset;
     (void) size;
@@ -8,7 +8,7 @@ static size_t read_null(fs_node_t* node, uint32_t offset, size_t size, uint8_t* 
 	return 0;
 }
 
-static size_t write_null(fs_node_t* node, uint32_t offset, size_t size, uint8_t* buf) {
+static ssize_t write_null(fs_node_t* node, off_t offset, size_t size, uint8_t* buf) {
     (void) node;
     (void) offset;
     (void) buf;
